Dragon---------2.5  *Modelos silenciosos y formularios artificiales   Esta funcionalidad le permite a dragon analizar las tablas que aun no se le ha creado un    Controlador, CRUD, o vista y lo crea automaticamente    Esto servira para los mantenimientos plains sean full automaticos    Tambien ayudara a que cuando se cree una vista, tabla no se necesite crear los modelos y controles/cruds    Estos modelos se pueden ver en la consola coomo "Silend Models"  *Mejoras de performances  *Ahora se podrá configurar una condición base al usuario para que el sistema se ponga modo cliente.     *Colocar mas tenue los iconos de la tabla     *El check de la tabla en el tema negro se nota más ancho     *Tomar en cuenta el drag en drog y el nuevo para edit cell     *Colocar mas pequeño los iconos dentro de la columnas  *Verificar los estilos de los filtros  *Dropdown modernos en los filtros  *En la pantalla de login verificar el estilo con resolución muy alta   Esto mismo pasa con el menú  *Title en el paréntesis de filtros, y en los conectores Y/O  *El boton de eliminar filtro quitarle el estilo forma botón  *Todos los botones tendrán solo iconos y su texto en el title  *Traducir el texto filter en el modal de filtros  *Traducir el texto remove filter en filtros  *En el grid de un solo registro ponerle secundario mas tenue   Las tablas con un solo registros se verán igual  *Alineado a la izquierda las acciones en la vista  *Los iconos de los controles bajarlos un poco, *Bajar el margen de los textos controles, *Darle estilo a el control de fecha rangos, *Textarea quitar la sangría, *Darle estilo al control de upload, *La ,inea del chozen múltiple no llega hasta el iconos, *Añadir el campo mascarilla en los campos personalizados, *Los choozen al agregar elemento borra los anteriores, *Pagina principal en la ,onfiguración, solo sandran los items del menu, *Colocar pagina por defecto en modulo de grupos  *Los archivos de loading que salgan igual  *Toda acción debe retroalimentar el usuario  *Se eliminó el cambio de color de los iconos por validación, para la igualdad  *titulos para los iconos de sort tan incomprendidos y solo se mostrarán cuando estén sorteados  *Los controles de fecha y montos ahora obedecen a la regla ISO de su idioma la aplicación manejará tipo de monedas por idioma

+++ admin/modules/BASE_storage.js
@@ -312,7 +312,11 @@ exports.defaultRequests = function (Model, params, folder) {
         params.modules.views.LoadEJSDragon(files, params, folder);
     });
     params.app.post('/api/st_list', function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             Model.all(req.query).then((data) => {
                 if (data.error !== false) res.send(data.error);
                 res.json(data);
@@ -322,7 +326,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.post(params.util.format('/api/%s/list', Model.tableName), function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
 
             if (req.query.limit === undefined)
                 req.query.limit = 10;
@@ -340,8 +348,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.get(params.util.format('/api/%s/all', Model.tableName), function (req, res) {
-        params.secure.check(req, res, function () {
-
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             if (req.query.limit === undefined)
                 req.query.limit = 10;
             if (req.query.page === undefined)
@@ -358,7 +369,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.get(params.util.format('/api/%s/get/:id', Model.tableName), function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             Model.find(req.params.id).then((data) => {
                 if (data.error !== false) res.send(data.error);
                 res.json(data);
@@ -368,7 +383,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.post('/api/' + Model.tableName + '/insert', function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             Model.insert(req.body).then((data) => {
                 if (data.error !== false) res.send(data.error);
                 res.json(data);
@@ -378,7 +397,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.post('/api/' + Model.tableName + '/insertID', function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             if (req.body.limit === undefined)
                 req.body.limit = 99999999;
             if (req.body.page === undefined)
@@ -395,7 +418,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.post('/api/' + Model.tableName + '/update/', function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             Model.update(req.body).then((data) => {
                 if (data.error !== false) res.send(data.error);
                 res.json(data);
@@ -405,7 +432,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.post('/api/' + Model.tableName + '/delete', function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             Model.delete(req.body).then((data) => {
                 if (data.error !== false) res.send(data.error);
                 res.json(data);
@@ -415,7 +446,11 @@ exports.defaultRequests = function (Model, params, folder) {
         });
     });
     params.app.post('/api/' + Model.tableName + '/truncate', function (req, res) {
-        params.secure.check(req, res, function () {
+        params.secure.check(req, res).then( function (token) {
+            if (!token.apptoken) {
+                res.json(token);
+                return;
+            }
             Model.truncate().then((data) => {
                 if (data.error !== false) res.send(data.error);
                 res.json(data);
